+++ packages/agent/src/observable.ts
@@ -2,19 +2,11 @@ import { AgentError } from './errors';
 
 export type ObserveFunction<T> = (data: T, ...rest: unknown[]) => void;
 
-export class Observable<T> extends Function {
+export class Observable<T> {
   observers: ObserveFunction<T>[];
 
   constructor() {
-    super();
     this.observers = [];
-    return new Proxy(this, {
-      apply: (target, _, args) => target.#call(args[0], ...args.slice(1)),
-    });
-  }
-
-  #call(message: T, ...rest: unknown[]) {
-    this.notify(message, ...rest);
   }
 
   subscribe(func: ObserveFunction<T>) {
@@ -44,11 +36,8 @@ export type AgentLog =
 export class ObservableLog extends Observable<AgentLog> {
   constructor() {
     super();
-    return new Proxy(this, {
-      apply: (target, _, args) => target.#call(args[0], ...args.slice(1)),
-    });
   }
-  log(message: string, ...rest: unknown[]) {
+  print(message: string, ...rest: unknown[]) {
     this.notify({ message, level: 'info' }, ...rest);
   }
   warn(message: string, ...rest: unknown[]) {
@@ -56,8 +45,5 @@ export class ObservableLog extends Observable<AgentLog> {
   }
   error(message: string, error: AgentError, ...rest: unknown[]) {
     this.notify({ message, level: 'error', error }, ...rest);
-  }
-  #call(message: string, ...rest: unknown[]) {
-    this.log(message, ...rest);
   }
 }