+++ docs/CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Changed
 
+- fix: ObservableLog no longer extends Function and class instance can no longer be called. Fixes an issue when running in a browser extension context.
+- feat!: ObservableLog's `log` method is renamed to `print` to avoind calling `log.log`.
 - chore: update management canister interface with latest bitcoin features
 - fix: publish script will correctly update the `package-lock.json` file with the correct dependencies when making a new release
 - chore: updates agent error response to read "Gateway returns error" instead of "Server"`